BB-190: Rename identifiers.jsx to reflect single component
